requirements.txt
Django>=1.6,<1.9
djangorestframework>=3.5,<3.7
isodate>=0.4.7,<0.6
